+++ src/prepare/layers.ts
@@ -1,7 +1,7 @@
-import type { I18nNuxtContext } from './context'
+import type { I18nNuxtContext } from '../context'
 import type { Nuxt } from '@nuxt/schema'
-import { applyLayerOptions } from './layers'
-import { filterLocales, mergeI18nModules } from './utils'
+import { applyLayerOptions } from '../layers'
+import { filterLocales, mergeI18nModules } from '../utils'
 
 export async function prepareLayers({ options }: I18nNuxtContext, nuxt: Nuxt) {
   applyLayerOptions(options, nuxt)
