+++ src/prepare/runtime-config.ts
@@ -1,6 +1,6 @@
-import type { I18nNuxtContext } from './context'
+import type { I18nNuxtContext } from '../context'
 import type { Nuxt } from '@nuxt/schema'
-import { DEFAULT_OPTIONS } from './constants'
+import { DEFAULT_OPTIONS } from '../constants'
 import { defu } from 'defu'
 
 export function prepareRuntimeConfig({ options }: I18nNuxtContext, nuxt: Nuxt) {
